Fix AS type conversion warning

+++ contracts/assemblyscript/erc20/assembly/lib-contract.ts
@@ -15,7 +15,7 @@ export function toBytes<T>(num: T, le: boolean = true): Uint8Array {
   // accept only integers and booleans
   if (isInteger<T>()) {
     const arr = new Uint8Array(sizeof<T>());
-    store<T>(arr.dataStart, le ? num : bswap(num));
+    store<T>(arr.dataStart as i32, le ? num : bswap(num));
     return arr;
   }
   assert(false);
